+++ uses-example/npm-test/node-js/demo.js
@@ -6,7 +6,7 @@ console.log('=== Running Devdoot JS from NPM Registry ===');
 devdoot.log('JS App started successfully.');
 
 // 2. Create isolated group logger
-const dbLog = devdoot.group('Database');
+const dbLog = devdoot.newGroup('Database');
 dbLog.log('Connected to JS DB Instance.');
 
 // 3. Info message