+++ uses-example/npm-test/typescript/demo.ts
@@ -6,7 +6,7 @@ console.log('=== Running Devdoot TS from NPM Registry ===');
 devdoot.log('TS App started successfully.');
 
 // 2. Create isolated group logger
-const dbLog = devdoot.group('Database');
+const dbLog = devdoot.newGroup('Database');
 dbLog.log('Connected to TS DB Instance.');
 
 // 3. Info message